stm32/l0: Add USART support (use F0 driver).
