+++ lib/stm32/l0/Makefile
@@ -44,6 +44,9 @@ OBJS		+= crs_common_all.o
 OBJS		+= flash.o
 OBJS		+= flash_common_l01.o
 OBJS		+= i2c_common_v2.o
+OBJS		+= usart.o
+OBJS		+= usart_common_l0f0.o
+OBJS		+= usart_common_v2.o
 
 OBJS            += usb.o usb_control.o usb_standard.o
 OBJS            += st_usbfs_core.o st_usbfs_v2.o
